convert to ns timer

+++ src/trusted/service_runtime/nacl_globals.c
@@ -43,8 +43,8 @@ void add_syscall_time(int sysnum, double call_time) {
 
 void print_execution_times(int sysnum) {
     if (lind_syscall_invoked_times[sysnum] == 0) return;
-    long average_time = lind_syscall_execution_time[sysnum]/lind_syscall_invoked_times[sysnum];
-    fprintf(stderr, "call number: %d, calls: %d, nsecs/call: %d \n", sysnum, lind_syscall_invoked_times[sysnum], average_time);
+    int average_time = lind_syscall_execution_time[sysnum]/lind_syscall_invoked_times[sysnum];
+    fprintf(stderr, "call number: %d, calls: %d, nsecs/call: %ld \n", sysnum, lind_syscall_invoked_times[sysnum], average_time);
 }
 
 long LindGetTime_ns(void) {
